parameters: dedupe assignment functions
