+++ cmd/parameters/setter.go
@@ -1,0 +1,58 @@
+package parameters
+
+import "context"
+
+// TODO: dedupe with cmds, we only really need to define a (provided, err) func like fromEnv is doing
+type providedFunc func(argument *Argument) (provided bool, err error)
+
+func setEach(ctx context.Context, providedFn providedFunc,
+	argsToSet ArgumentList,
+	inputErrors <-chan error) (ArgumentList, <-chan error) {
+	var (
+		unsetArgs = make(chan *Argument, cap(argsToSet))
+		errors    = make(chan error, cap(inputErrors))
+	)
+	go func() {
+		defer close(unsetArgs)
+		defer close(errors)
+		for argsToSet != nil ||
+			inputErrors != nil {
+			select {
+			case argument, ok := <-argsToSet:
+				if !ok {
+					argsToSet = nil
+					continue
+				}
+				provided, err := providedFn(argument)
+				if err != nil {
+					select {
+					case errors <- err:
+					case <-ctx.Done():
+					}
+				}
+				if provided {
+					continue
+				}
+				select { // Relay parameter to next source.
+				case unsetArgs <- argument:
+				case <-ctx.Done():
+				}
+			case err, ok := <-inputErrors:
+				if !ok {
+					inputErrors = nil
+					continue
+				}
+				// If we encounter an input error,
+				// relay it and keep processing.
+				// The caller may decide to cancel or not afterwards.
+				select {
+				case errors <- err:
+				case <-ctx.Done():
+				}
+			case <-ctx.Done():
+				return
+			}
+		}
+	}()
+	return unsetArgs, errors
+}
